refact(Braintree): Move interfaces to paymentProvider/types.ts
