+++ server/paymentProviders/types.ts
@@ -1,7 +1,5 @@
 import models from '../models';
 
-import { PaymentProviderService } from './PaymentProviderService';
-
 export interface PaymentProvider {
   /**
    * Triggers the payment for this order and updates it accordingly
@@ -12,4 +10,18 @@ export interface PaymentProvider {
    * The different types of payment methods supported by this provider
    */
   types: Record<string, PaymentProviderService>;
+}
+
+export interface PaymentProviderService {
+  /**
+   * Describes the features implemented by this payment method
+   */
+  features: {
+    recurring: boolean;
+  };
+
+  /**
+   * Triggers the payment for this order and updates it accordingly
+   */
+  processOrder(order: typeof models.Order): Promise<typeof models.Transaction>;
 }
